+++ CHANGELOG.md
@@ -79,6 +79,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     provided in a resource can be converted to a `[System.TimeSpan]` and
     optionally evaluates so it is not below a minium value or over a maximum
     value.
+  - Added new private function `Get-ClassName` that returns the class name
+    or optionally an array with the class name and all inherited base class
+    named.
+  - Added new private function `Get-LocalizedDataRecursive` that gathers
+    all localization strings from an array of class names. This can be used
+    in classes to be able to inherit localization strings from one or more
+    base class. If a localization string key exist in a parent class's
+    localization string file it will override the localization string key
+    in any base class.
   - Added `prefix.ps1` that is used to import dependent modules like _DscResource.Common_.
   - Added new resource
     - _DnsServerScavenging_ - resource to enforce scavenging settings ([issue #189](https://github.com/dsccommunity/xDnsServer/issues/189)).
@@ -144,6 +153,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - Fixed the sections in the GitHub issue and pull request templates to
     have a bit higher font size. This makes it easier to distinguish the
     section headers from the text.
+- ResourceBase
+  - Added support for inherit localization strings and also able to override
+    a localization string that exist in a base class.
 - DnsRecordBase
   - Changed class to inherit properties from 'ResourcePropertiesBase`.
 - xDnsRecordSrv